Update Invoke-CMDownloadBIOSPackage.ps1
Troubleshooting: device has some UEFI updates, but none is selected instead of the last update. It's a casting error at selection by all lines with: $PackageList = $PackageList | Sort-Object -Property SourceDate -Descending | Select-Object -First 1

Solution: Create an array object and not a PSCustom object, so that the method count is present.